requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# ML and Data Processing
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
pandas==2.1.4
numpy==1.25.2
shap==0.44.0

# Document Processing
PyPDF2==3.0.1
openpyxl==3.1.2
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6

# LLM and Embeddings
sentence-transformers==2.2.2
chromadb==0.4.18
ragas==0.0.22

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic-settings==2.1.0 